Extract formatting cache to dedicated module
Extracted cache-related functionality from format.ts (2241→2168 lines) into
src/cli/src/modules/formatting/ following established CLI module pattern.

Moved functions: trimFormattingCache, getFormattingCacheEntry,
storeFormattingCacheEntry, estimateFormattingCacheBytes,
createFormattingCacheKey, plus cache Map and related helpers.

Updated format.ts to import from new module. Updated test helpers to
use exported cache functions. Removed unused createHash import.
