+++ src/cli/src/modules/formatting/index.ts
@@ -1,0 +1,10 @@
+export {
+    clearFormattingCache,
+    createFormattingCacheKey,
+    estimateFormattingCacheBytes,
+    getFormattingCacheEntry,
+    getFormattingCacheKeys,
+    getFormattingCacheStats,
+    storeFormattingCacheEntry,
+    trimFormattingCache
+} from "./cache.js";
